+++ chain/network/src/routing/mod.rs
@@ -1,8 +1,8 @@
-mod cache;
 pub mod codec;
 pub mod edge_verifier_actor;
 pub mod ibf;
 pub mod ibf_peer_set;
 pub mod ibf_set;
+mod route_back_cache;
 pub mod routing;
 pub mod routing_table_actor;
